requirements.txt
requests>=2.31.0
yara-python>=4.3.0
tkinterdnd2>=0.3.0
pyinstaller>=6.0.0
